chore: remove workspace
Removes workspace and just have a single module for simplicity.

I'd never used workspaces before, and was hoping I could leverage them
to keep the dependencies for the examples separate from the dependencies
of the library. So far though that has just been a source of headaches,
and I'm not evene sure if this is an intended use-case for workspaces,
so I figured it was better to just stick with the simple layout for now
while in the early stages of development.