fix: remove auto-cleanup of old test databases
As I wrote it, this is not safe to have on by default if there are
multiple packages being tested which make use of `pgtest`. Mainly, the
auto-cleanup can end up trying to delete test databases still in use by
another `pgtest.Supervisor` instanve from another package.

There should be a way to make this safe, likely using postgres locks,
but for now just commenting out.