chore: add README, better simple example.
Adds a README providing a high level overview of this library with a
straightforward quick start. Added a new example as well since I
realized that although the examples I added already were good at
demonstrating the use of `pgtest` in something similar to a real world
application, they were too complex to really work as an introduction.