fix: handle other possible error code for duplicate db
Treat errors with code 23505 (unique_violation) the same as those with
code 42P04 (duplicate_database) when creating test database.

Unfortunately I didn't find clear documentation explaining this, but I
have encountered both of these error codes when attempting to create a
database with the same name as an existing database. I'm assuming this
is just something that is different across different postgresql
versions, and while it would be nice to have confirmation we can just
treat these error codes as equivalent in this context.